+++ crmsh/user_of_host.py
@@ -47,7 +47,7 @@ class UserOfHost:
     def user_pair_for_ssh(self, host: str) -> typing.Tuple[str, str]:
         """Return (local_user, remote_user) pair for ssh connection"""
         if config.core.no_ssh:
-            from .utils import NoSSHError
+            from .network_utils import NoSSHError
             raise NoSSHError(constants.NO_SSH_ERROR_MSG)
         local_user = None
         remote_user = None
