+++ data-manifest
@@ -201,6 +201,7 @@ test/unittests/test_gv.py
 test/unittests/test_handles.py
 test/unittests/test_lock.py
 test/unittests/test_migration.py
+test/unittests/test_network_utils.py
 test/unittests/test_objset.py
 test/unittests/test_parallax.py
 test/unittests/test_parse.py
