+++ test/unittests/test_utils.py
@@ -4,7 +4,6 @@
 # unit tests for utils.py
 
 import os
-import socket
 import re
 import unittest
 import pytest
@@ -26,7 +25,7 @@ def test_package_is_installed_local(mock_run):
     mock_run.assert_called_once_with("rpm -q --quiet crmsh")
 
 
-@mock.patch('crmsh.xmlutil.CrmMonXmlParser')
+@mock.patch("crmsh.utils.xmlutil.CrmMonXmlParser")
 def test_get_nodeid_from_name_remote(mock_parser):
     mock_parser_inst = mock.Mock()
     mock_parser.return_value = mock_parser_inst
@@ -36,7 +35,7 @@ def test_get_nodeid_from_name_remote(mock_parser):
     mock_parser_inst.is_node_remote.assert_called_once_with("node1")
 
 
-@mock.patch('crmsh.xmlutil.CrmMonXmlParser')
+@mock.patch("crmsh.utils.xmlutil.CrmMonXmlParser")
 def test_get_nodeid_from_name(mock_parser):
     mock_parser_inst = mock.Mock()
     mock_parser.return_value = mock_parser_inst
@@ -48,15 +47,98 @@ def test_get_nodeid_from_name(mock_parser):
     mock_parser_inst.get_node_id_from_name.assert_called_once_with("node1")
 
 
-@mock.patch('crmsh.sh.LocalShell.get_rc_and_error')
-def test_check_ssh_passwd_need(mock_run):
-    mock_run.return_value = (1, 'foo')
-    res = utils.check_ssh_passwd_need("bob", "alice", "node1")
-    assert res is True
-    mock_run.assert_called_once_with(
-        "bob",
-        "ssh -o StrictHostKeyChecking=no -o EscapeChar=none -o ConnectTimeout=15 -T -o Batchmode=yes alice@node1 true",
-    )
+
+@mock.patch("crmsh.sh.ShellUtils.get_stdout")
+def test_get_nodeinfo_from_cmaptool_return_none(mock_get_stdout):
+    mock_get_stdout.return_value = (1, None)
+    assert bool(utils.get_nodeinfo_from_cmaptool()) is False
+    mock_get_stdout.assert_called_once_with("corosync-cmapctl -b runtime.members")
+
+
+@mock.patch("crmsh.utils.re.findall")
+@mock.patch("crmsh.utils.re.search")
+@mock.patch("crmsh.sh.ShellUtils.get_stdout")
+def test_get_nodeinfo_from_cmaptool(mock_get_stdout, mock_search, mock_findall):
+    mock_get_stdout.return_value = (0, 'runtime.members.1.ip (str) = r(0) ip(192.168.43.129)\nruntime.members.2.ip (str) = r(0) ip(192.168.43.128)')
+    match_inst1 = mock.Mock()
+    match_inst2 = mock.Mock()
+    mock_search.side_effect = [match_inst1, match_inst2]
+    match_inst1.group.return_value = '1'
+    match_inst2.group.return_value = '2'
+    mock_findall.side_effect = [["192.168.43.129"], ["192.168.43.128"]]
+
+    result = utils.get_nodeinfo_from_cmaptool()
+    assert result['1'] == ["192.168.43.129"]
+    assert result['2'] == ["192.168.43.128"]
+
+    mock_get_stdout.assert_called_once_with("corosync-cmapctl -b runtime.members")
+    mock_search.assert_has_calls([
+        mock.call(r'members\.(.*)\.ip', 'runtime.members.1.ip (str) = r(0) ip(192.168.43.129)'),
+        mock.call(r'members\.(.*)\.ip', 'runtime.members.2.ip (str) = r(0) ip(192.168.43.128)')
+    ])
+    match_inst1.group.assert_called_once_with(1)
+    match_inst2.group.assert_called_once_with(1)
+    mock_findall.assert_has_calls([
+        mock.call(r'[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}', 'runtime.members.1.ip (str) = r(0) ip(192.168.43.129)'),
+        mock.call(r'[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}', 'runtime.members.2.ip (str) = r(0) ip(192.168.43.128)')
+    ])
+
+@mock.patch("crmsh.utils.get_nodeinfo_from_cmaptool")
+@mock.patch("crmsh.service_manager.ServiceManager.service_is_active")
+def test_valid_nodeid_false_service_not_active(mock_is_active, mock_nodeinfo):
+    mock_is_active.return_value = False
+    assert utils.valid_nodeid("3") is False
+    mock_is_active.assert_called_once_with('corosync.service')
+    mock_nodeinfo.assert_not_called()
+
+@mock.patch("crmsh.utils.get_nodeinfo_from_cmaptool")
+@mock.patch("crmsh.service_manager.ServiceManager.service_is_active")
+def test_valid_nodeid_false(mock_is_active, mock_nodeinfo):
+    mock_is_active.return_value = True
+    mock_nodeinfo.return_value = {'1': ["10.10.10.1"], "2": ["20.20.20.2"]}
+    assert utils.valid_nodeid("3") is False
+    mock_is_active.assert_called_once_with('corosync.service')
+    mock_nodeinfo.assert_called_once_with()
+
+@mock.patch("crmsh.utils.get_nodeinfo_from_cmaptool")
+@mock.patch("crmsh.service_manager.ServiceManager.service_is_active")
+def test_valid_nodeid_true(mock_is_active, mock_nodeinfo):
+    mock_is_active.return_value = True
+    mock_nodeinfo.return_value = {'1': ["10.10.10.1"], "2": ["20.20.20.2"]}
+    assert utils.valid_nodeid("2") is True
+    mock_is_active.assert_called_once_with('corosync.service')
+    mock_nodeinfo.assert_called_once_with()
+
+
+@mock.patch("crmsh.utils.get_nodeid_from_name")
+def test_get_iplist_from_name_no_nodeid(mock_get_nodeid):
+    mock_get_nodeid.return_value = None
+    res = utils.get_iplist_from_name("test")
+    assert res == []
+    mock_get_nodeid.assert_called_once_with("test")
+
+
+@mock.patch("crmsh.utils.get_nodeinfo_from_cmaptool")
+@mock.patch("crmsh.utils.get_nodeid_from_name")
+def test_get_iplist_from_name_no_nodeinfo(mock_get_nodeid, mock_get_nodeinfo):
+    mock_get_nodeid.return_value = "1"
+    mock_get_nodeinfo.return_value = None
+    res = utils.get_iplist_from_name("test")
+    assert res == []
+    mock_get_nodeid.assert_called_once_with("test")
+    mock_get_nodeinfo.assert_called_once_with()
+
+
+@mock.patch("crmsh.utils.get_nodeinfo_from_cmaptool")
+@mock.patch("crmsh.utils.get_nodeid_from_name")
+def test_get_iplist_from_name(mock_get_nodeid, mock_get_nodeinfo):
+    mock_get_nodeid.return_value = "1"
+    mock_get_nodeinfo.return_value = {"1": ["10.10.10.1"], "2": ["10.10.10.2"]}
+    res = utils.get_iplist_from_name("test")
+    assert res == ["10.10.10.1"]
+    mock_get_nodeid.assert_called_once_with("test")
+    mock_get_nodeinfo.assert_called_once_with()
+
 
 
 @mock.patch('crmsh.utils.list_cluster_nodes')
@@ -248,112 +330,6 @@ def test_sysconfig_set_bsc1145823():
     sc = utils.parse_sysconfig(fname)
     assert (sc.get("age") == "100")
 
-@mock.patch("socket.getaddrinfo")
-@mock.patch("socket.socket")
-@mock.patch("selectors.DefaultSelector")
-@mock.patch("time.sleep")
-def test_check_port_open_false(mock_sleep, mock_selector_cls, mock_socket, mock_getaddrinfo):
-    sock_inst = mock.Mock()
-    mock_socket.return_value = sock_inst
-    sock_inst.connect_ex.return_value = 1
-    mock_getaddrinfo.return_value = [(socket.AF_INET, socket.SOCK_STREAM, 6, "", ("127.0.0.1", 22))]
-
-    mock_selector = mock.Mock()
-    mock_selector_cls.return_value = mock_selector
-    mock_selector.select.return_value = []
-
-    assert utils.check_port_open("localhost", 22) is False
-
-    assert mock_socket.call_count == 3
-    assert sock_inst.connect_ex.call_count == 3
-    assert mock_selector_cls.call_count == 3
-    assert mock_selector.select.call_count == 3
-    assert mock_sleep.call_count == 2
-
-@mock.patch("socket.getaddrinfo")
-@mock.patch("socket.socket")
-@mock.patch("selectors.DefaultSelector")
-def test_check_port_open_true(mock_selector_cls, mock_socket, mock_getaddrinfo):
-    sock_inst = mock.Mock()
-    mock_socket.return_value = sock_inst
-    sock_inst.connect_ex.return_value = 0
-    mock_getaddrinfo.return_value = [(socket.AF_INET, socket.SOCK_STREAM, 6, "", ("127.0.0.1", 22))]
-
-    mock_selector = mock.Mock()
-    mock_selector_cls.return_value = mock_selector
-
-    assert utils.check_port_open("localhost", 22) is True
-
-    assert mock_socket.call_count == 1
-    assert sock_inst.connect_ex.call_count == 1
-    assert mock_selector_cls.call_count == 1
-
-def test_valid_port():
-    assert utils.valid_port(1) is False
-    assert utils.valid_port(10000000) is False
-    assert utils.valid_port(1234) is True
-
-
-@mock.patch("crmsh.sh.ShellUtils.get_stdout")
-def test_get_nodeinfo_from_cmaptool_return_none(mock_get_stdout):
-    mock_get_stdout.return_value = (1, None)
-    assert bool(utils.get_nodeinfo_from_cmaptool()) is False
-    mock_get_stdout.assert_called_once_with("corosync-cmapctl -b runtime.members")
-
-
-@mock.patch("re.findall")
-@mock.patch("re.search")
-@mock.patch("crmsh.sh.ShellUtils.get_stdout")
-def test_get_nodeinfo_from_cmaptool(mock_get_stdout, mock_search, mock_findall):
-    mock_get_stdout.return_value = (0, 'runtime.members.1.ip (str) = r(0) ip(192.168.43.129)\nruntime.members.2.ip (str) = r(0) ip(192.168.43.128)')
-    match_inst1 = mock.Mock()
-    match_inst2 = mock.Mock()
-    mock_search.side_effect = [match_inst1, match_inst2]
-    match_inst1.group.return_value = '1'
-    match_inst2.group.return_value = '2'
-    mock_findall.side_effect = [["192.168.43.129"], ["192.168.43.128"]]
-
-    result = utils.get_nodeinfo_from_cmaptool()
-    assert result['1'] == ["192.168.43.129"]
-    assert result['2'] == ["192.168.43.128"]
-
-    mock_get_stdout.assert_called_once_with("corosync-cmapctl -b runtime.members")
-    mock_search.assert_has_calls([
-        mock.call(r'members\.(.*)\.ip', 'runtime.members.1.ip (str) = r(0) ip(192.168.43.129)'),
-        mock.call(r'members\.(.*)\.ip', 'runtime.members.2.ip (str) = r(0) ip(192.168.43.128)')
-    ])
-    match_inst1.group.assert_called_once_with(1)
-    match_inst2.group.assert_called_once_with(1)
-    mock_findall.assert_has_calls([
-        mock.call(r'[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}', 'runtime.members.1.ip (str) = r(0) ip(192.168.43.129)'),
-        mock.call(r'[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}', 'runtime.members.2.ip (str) = r(0) ip(192.168.43.128)')
-    ])
-
-@mock.patch("crmsh.utils.get_nodeinfo_from_cmaptool")
-@mock.patch("crmsh.service_manager.ServiceManager.service_is_active")
-def test_valid_nodeid_false_service_not_active(mock_is_active, mock_nodeinfo):
-    mock_is_active.return_value = False
-    assert utils.valid_nodeid("3") is False
-    mock_is_active.assert_called_once_with('corosync.service')
-    mock_nodeinfo.assert_not_called()
-
-@mock.patch("crmsh.utils.get_nodeinfo_from_cmaptool")
-@mock.patch("crmsh.service_manager.ServiceManager.service_is_active")
-def test_valid_nodeid_false(mock_is_active, mock_nodeinfo):
-    mock_is_active.return_value = True
-    mock_nodeinfo.return_value = {'1': ["10.10.10.1"], "2": ["20.20.20.2"]}
-    assert utils.valid_nodeid("3") is False
-    mock_is_active.assert_called_once_with('corosync.service')
-    mock_nodeinfo.assert_called_once_with()
-
-@mock.patch("crmsh.utils.get_nodeinfo_from_cmaptool")
-@mock.patch("crmsh.service_manager.ServiceManager.service_is_active")
-def test_valid_nodeid_true(mock_is_active, mock_nodeinfo):
-    mock_is_active.return_value = True
-    mock_nodeinfo.return_value = {'1': ["10.10.10.1"], "2": ["20.20.20.2"]}
-    assert utils.valid_nodeid("2") is True
-    mock_is_active.assert_called_once_with('corosync.service')
-    mock_nodeinfo.assert_called_once_with()
 
 @mock.patch('crmsh.sh.ClusterShell.get_stdout_or_raise_error')
 def test_detect_aws_false(mock_run):
@@ -470,264 +446,6 @@ def test_detect_cloud_gcp(mock_is_program, mock_aws, mock_azure, mock_gcp):
     mock_gcp.assert_called_once_with()
 
 
-class TestIP(unittest.TestCase):
-    """
-    Unitary tests for class utils.IP
-    """
-    @classmethod
-    def setUpClass(cls):
-        """
-        Global setUp.
-        """
-
-    def setUp(self):
-        """
-        Test setUp.
-        """
-        self.ip_inst = utils.IP("10.10.10.1")
-
-    def tearDown(self):
-        """
-        Test tearDown.
-        """
-
-    @classmethod
-    def tearDownClass(cls):
-        """
-        Global tearDown.
-        """
-
-    @mock.patch('ipaddress.ip_address')
-    def test_ip_address(self, mock_ip_address):
-        mock_ip_address_inst = mock.Mock()
-        mock_ip_address.return_value = mock_ip_address_inst
-        self.ip_inst.ip_address
-        mock_ip_address.assert_called_once_with("10.10.10.1")
-
-    @mock.patch('crmsh.utils.IP.ip_address', new_callable=mock.PropertyMock)
-    def test_version(self, mock_ip_address):
-        mock_ip_address_inst = mock.Mock(version=4)
-        mock_ip_address.return_value = mock_ip_address_inst
-        res = self.ip_inst.version
-        self.assertEqual(res, mock_ip_address_inst.version)
-        mock_ip_address.assert_called_once_with()
-
-    @mock.patch('crmsh.utils.IP.ip_address', new_callable=mock.PropertyMock)
-    def test_is_mcast(self, mock_ip_address):
-        mock_ip_address_inst = mock.Mock(is_multicast=False)
-        mock_ip_address.return_value = mock_ip_address_inst
-        res = utils.IP.is_mcast("10.10.10.1")
-        self.assertEqual(res, False)
-        mock_ip_address.assert_called_once_with()
-
-    @mock.patch('crmsh.utils.IP.version', new_callable=mock.PropertyMock)
-    def test_is_ipv6(self, mock_version):
-        mock_version.return_value = 4
-        res = utils.IP.is_ipv6("10.10.10.1")
-        self.assertEqual(res, False)
-        mock_version.assert_called_once_with()
-
-    @mock.patch('crmsh.utils.IP.ip_address', new_callable=mock.PropertyMock)
-    def test_is_loopback(self, mock_ip_address):
-        mock_ip_address_inst = mock.Mock(is_loopback=False)
-        mock_ip_address.return_value = mock_ip_address_inst
-        res = self.ip_inst.is_loopback
-        self.assertEqual(res, mock_ip_address_inst.is_loopback)
-        mock_ip_address.assert_called_once_with()
-
-
-class TestInterface(unittest.TestCase):
-    """
-    Unitary tests for class utils.Interface
-    """
-    @classmethod
-    def setUpClass(cls):
-        """
-        Global setUp.
-        """
-
-    def setUp(self):
-        """
-        Test setUp.
-        """
-        self.interface = utils.Interface("10.10.10.123/24")
-
-    def tearDown(self):
-        """
-        Test tearDown.
-        """
-
-    @classmethod
-    def tearDownClass(cls):
-        """
-        Global tearDown.
-        """
-
-    def test_ip_with_mask(self):
-        assert self.interface.ip_with_mask == "10.10.10.123/24"
-
-    @mock.patch('ipaddress.ip_interface')
-    def test_ip_interface(self, mock_ip_interface):
-        mock_ip_interface_inst = mock.Mock()
-        mock_ip_interface.return_value = mock_ip_interface_inst
-        self.interface.ip_interface
-        mock_ip_interface.assert_called_once_with("10.10.10.123/24")
-
-    @mock.patch('crmsh.utils.Interface.ip_interface', new_callable=mock.PropertyMock)
-    def test_network(self, mock_ip_interface):
-        mock_ip_interface_inst = mock.Mock()
-        mock_ip_interface.return_value = mock_ip_interface_inst
-        mock_ip_interface_inst.network = mock.Mock(network_address="10.10.10.0")
-        assert self.interface.network == "10.10.10.0"
-        mock_ip_interface.assert_called_once_with()
-
-
-class TestInterfacesInfo(unittest.TestCase):
-    """
-    Unitary tests for class utils.InterfacesInfo
-    """
-
-    network_output_error = """1: lo    inet 127.0.0.1/8 scope host lo\\       valid_lft forever preferred_lft forever
-2: enp1s0    inet 192.168.122.241/24 brd 192.168.122.255 scope global enp1s0
-61: tun0    inet 10.163.45.46 peer 10.163.45.45/32 scope global tun0"""
-
-    @classmethod
-    def setUpClass(cls):
-        """
-        Global setUp.
-        """
-
-    def setUp(self):
-        """
-        Test setUp.
-        """
-        self.interfaces_info = utils.InterfacesInfo()
-        self.interfaces_info_with_custom_nic = utils.InterfacesInfo(custom_nic_addr_list=['eth1'])
-        self.interfaces_info_with_wrong_nic = utils.InterfacesInfo(custom_nic_addr_list=['eth7'])
-        self.interfaces_info_fake = utils.InterfacesInfo()
-        self.interfaces_info_fake._nic_info_dict = {
-                "eth0": [mock.Mock(ip="10.10.10.1", network="10.10.10.0"), mock.Mock(ip="10.10.10.2", network="10.10.10.0")],
-                "eth1": [mock.Mock(ip="20.20.20.1", network="20.20.20.0")]
-                }
-        self.interfaces_info_fake._default_nic_list = ["eth7"]
-
-    def tearDown(self):
-        """
-        Test tearDown.
-        """
-
-    @classmethod
-    def tearDownClass(cls):
-        """
-        Global tearDown.
-        """
-
-    @mock.patch('crmsh.sh.ShellUtils.get_stdout_stderr')
-    def test_get_interfaces_info_no_address(self, mock_run):
-        only_lo = "1: lo    inet 127.0.0.1/8 scope host lo\\       valid_lft forever preferred_lft forever"
-        mock_run.return_value = (0, only_lo, None)
-        with self.assertRaises(ValueError) as err:
-            self.interfaces_info.get_interfaces_info()
-        self.assertEqual("No address configured", str(err.exception))
-        mock_run.assert_called_once_with("ip -4 -o addr show")
-
-    def test_nic_list(self):
-        res = self.interfaces_info_fake.nic_list
-        self.assertEqual(res, ["eth0", "eth1"])
- 
-    def test_interface_list(self):
-        res = self.interfaces_info_fake.interface_list
-        assert len(res) == 3
-
-    @mock.patch('crmsh.utils.InterfacesInfo.interface_list', new_callable=mock.PropertyMock)
-    def test_ip_list(self, mock_interface_list):
-        mock_interface_list.return_value = [
-                mock.Mock(ip="10.10.10.1"),
-                mock.Mock(ip="10.10.10.2")
-                ]
-        res = self.interfaces_info_fake.ip_list
-        self.assertEqual(res, ["10.10.10.1", "10.10.10.2"])
-        mock_interface_list.assert_called_once_with()
-
-    @mock.patch('crmsh.utils.InterfacesInfo.ip_list', new_callable=mock.PropertyMock)
-    @mock.patch('crmsh.utils.InterfacesInfo.get_interfaces_info')
-    def test_get_local_ip_list(self, mock_get_info, mock_ip_list):
-        mock_ip_list.return_value = ["10.10.10.1", "10.10.10.2"]
-        res = utils.InterfacesInfo.get_local_ip_list(False)
-        self.assertEqual(res, mock_ip_list.return_value)
-        mock_get_info.assert_called_once_with()
-        mock_ip_list.assert_called_once_with()
-
-    @mock.patch('crmsh.utils.InterfacesInfo.ip_list', new_callable=mock.PropertyMock)
-    @mock.patch('crmsh.utils.IP.is_ipv6')
-    @mock.patch('crmsh.utils.InterfacesInfo.get_interfaces_info')
-    def test_ip_in_local(self, mock_get_info, mock_is_ipv6, mock_ip_list):
-        mock_is_ipv6.return_value = False
-        mock_ip_list.return_value = ["10.10.10.1", "10.10.10.2"]
-        res = utils.InterfacesInfo.ip_in_local("10.10.10.1")
-        assert res is True
-        mock_get_info.assert_called_once_with()
-        mock_ip_list.assert_called_once_with()
-        mock_is_ipv6.assert_called_once_with("10.10.10.1")
-
-    @mock.patch('crmsh.utils.InterfacesInfo.interface_list', new_callable=mock.PropertyMock)
-    def test_network_list(self, mock_interface_list):
-        mock_interface_list.return_value = [
-                mock.Mock(network="10.10.10.0"),
-                mock.Mock(network="20.20.20.0")
-                ]
-        res = self.interfaces_info.network_list
-        self.assertEqual(res, list(set(["10.10.10.0", "20.20.20.0"])))
-        mock_interface_list.assert_called_once_with()
-
-    def test_nic_first_ip(self):
-        res = self.interfaces_info_fake.nic_first_ip("eth0")
-        self.assertEqual(res, "10.10.10.1")
-
-    @mock.patch('crmsh.sh.cluster_shell')
-    def test_get_default_nic_from_route(self, mock_run):
-        output = """default via 192.168.122.1 dev eth8 proto dhcp 
-        10.10.10.0/24 dev eth1 proto kernel scope link src 10.10.10.51 
-        20.20.20.0/24 dev eth2 proto kernel scope link src 20.20.20.51 
-        192.168.122.0/24 dev eth8 proto kernel scope link src 192.168.122.120"""
-        mock_run_inst = mock.Mock()
-        mock_run.return_value = mock_run_inst
-        mock_run_inst.get_stdout_or_raise_error.return_value = output
-
-        res = self.interfaces_info.get_default_nic_from_route()
-        self.assertEqual(res, "eth8")
-
-        mock_run_inst.get_stdout_or_raise_error.assert_called_once_with("ip -o route show")
-
-
-@mock.patch("crmsh.utils.get_nodeid_from_name")
-def test_get_iplist_from_name_no_nodeid(mock_get_nodeid):
-    mock_get_nodeid.return_value = None
-    res = utils.get_iplist_from_name("test")
-    assert res == []
-    mock_get_nodeid.assert_called_once_with("test")
-
-
-@mock.patch("crmsh.utils.get_nodeinfo_from_cmaptool")
-@mock.patch("crmsh.utils.get_nodeid_from_name")
-def test_get_iplist_from_name_no_nodeinfo(mock_get_nodeid, mock_get_nodeinfo):
-    mock_get_nodeid.return_value = "1"
-    mock_get_nodeinfo.return_value = None
-    res = utils.get_iplist_from_name("test")
-    assert res == []
-    mock_get_nodeid.assert_called_once_with("test")
-    mock_get_nodeinfo.assert_called_once_with()
-
-
-@mock.patch("crmsh.utils.get_nodeinfo_from_cmaptool")
-@mock.patch("crmsh.utils.get_nodeid_from_name")
-def test_get_iplist_from_name(mock_get_nodeid, mock_get_nodeinfo):
-    mock_get_nodeid.return_value = "1"
-    mock_get_nodeinfo.return_value = {"1": ["10.10.10.1"], "2": ["10.10.10.2"]}
-    res = utils.get_iplist_from_name("test")
-    assert res == ["10.10.10.1"]
-    mock_get_nodeid.assert_called_once_with("test")
-    mock_get_nodeinfo.assert_called_once_with()
 
 
 def test_calculate_quorate_status():
@@ -798,35 +516,6 @@ fencing-sbd
     mock_diskless.assert_called_once_with()
 
 
-@mock.patch('crmsh.utils.ssh_port_reachable_check')
-@mock.patch('crmsh.xmlutil.CrmMonXmlParser')
-def test_check_all_nodes_reachable_dead_nodes(mock_xml, mock_reachable):
-    mock_xml_inst = mock.Mock()
-    mock_xml.return_value = mock_xml_inst
-    mock_xml_inst.not_connected.return_value = False
-    mock_xml_inst.get_node_list.side_effect = [["node1"], ["node2"]]
-    mock_reachable.side_effect = ValueError
-
-    with pytest.raises(utils.DeadNodeError) as err:
-        utils.check_all_nodes_reachable("testing")
-    assert err.value.summary.dead_nodes == ["node2"]
-
-
-@mock.patch('crmsh.utils.check_ssh_passwd_need')
-@mock.patch('crmsh.user_of_host.UserOfHost.instance')
-@mock.patch('crmsh.utils.ssh_port_reachable_check')
-@mock.patch('crmsh.xmlutil.CrmMonXmlParser')
-def test_check_all_nodes_reachable(mock_xml, mock_reachable, mock_user_of_host, mock_check_passwd):
-    mock_xml_inst = mock.Mock()
-    mock_xml.return_value = mock_xml_inst
-    mock_xml_inst.not_connected.return_value = False
-    mock_xml_inst.get_node_list.side_effect = [["node1"], []]
-    mock_user_of_host_inst = mock.Mock()
-    mock_user_of_host.return_value = mock_user_of_host_inst
-    mock_user_of_host_inst.user_pair_for_ssh = mock.Mock(return_value=("root", "root"))
-    mock_check_passwd.return_value = False
-    utils.check_all_nodes_reachable("testing")
-    mock_reachable.assert_called_once_with("node1")
 
 
 @mock.patch('crmsh.sh.ShellUtils.get_stdout_stderr')
@@ -1157,17 +846,6 @@ def test_is_dc_idle(mock_dc, mock_shell):
     assert utils.is_dc_idle() is True
 
 
-@mock.patch('logging.Logger.warning')
-@mock.patch('crmsh.utils.ssh_port_reachable_check')
-def test_get_reachable_node_list(mock_reachable, mock_warn):
-    mock_reachable.side_effect = [False, True, ValueError("error for node3")]
-    assert utils.get_reachable_node_list(["node1", "node2", "node3"]) == ["node2"]
-    mock_warn.assert_called_once_with("error for node3")
-    mock_reachable.assert_has_calls([
-        mock.call("node1"),
-        mock.call("node2"),
-        mock.call("node3")
-    ])
 
 
 def test_verify_result():
@@ -1223,7 +901,7 @@ def test_validate_and_get_reachable_nodes_return_local(mock_list_nodes, mock_thi
     assert res == ["node1"]
 
 
-@mock.patch('crmsh.utils.get_reachable_node_list')
+@mock.patch('crmsh.network_utils.get_reachable_node_list')
 @mock.patch('crmsh.utils.this_node')
 @mock.patch('crmsh.utils.get_address_list_from_corosync_conf')
 @mock.patch('crmsh.utils.list_cluster_nodes')
@@ -1239,7 +917,7 @@ def test_validate_and_get_reachable_nodes_no_cib(mock_list_nodes, mock_get_addre
 @mock.patch('logging.Logger.error')
 @mock.patch('crmsh.xmlutil.CrmMonXmlParser')
 @mock.patch('crmsh.sh.cluster_shell')
-@mock.patch('crmsh.utils.get_reachable_node_list')
+@mock.patch('crmsh.network_utils.get_reachable_node_list')
 @mock.patch('crmsh.utils.this_node')
 @mock.patch('crmsh.utils.list_cluster_nodes')
 def test_validate_and_get_reachable_nodes(mock_list_nodes, mock_this_node, mock_get_reachable, mock_shell, mock_xml, mock_error):
